feat: update api

+++ larksuite-oapi/src/main/java/com/lark/oapi/service/hire/v1/model/RegistrationBasicInfo.java
@@ -39,7 +39,7 @@ public class RegistrationBasicInfo {
      * <p> 示例值：1618494330932
      */
     @SerializedName("registration_time")
-    private Integer registrationTime;
+    private Long registrationTime;
     /**
      * 下载链接
      * <p> 示例值：https://hire.feishu.cn/hire/file/blob/...token.../
@@ -92,11 +92,11 @@ public class RegistrationBasicInfo {
         this.id = id;
     }
 
-    public Integer getRegistrationTime() {
+    public Long getRegistrationTime() {
         return this.registrationTime;
     }
 
-    public void setRegistrationTime(Integer registrationTime) {
+    public void setRegistrationTime(Long registrationTime) {
         this.registrationTime = registrationTime;
     }
 
@@ -126,7 +126,7 @@ public class RegistrationBasicInfo {
          * 创建时间
          * <p> 示例值：1618494330932
          */
-        private Integer registrationTime;
+        private Long registrationTime;
         /**
          * 下载链接
          * <p> 示例值：https://hire.feishu.cn/hire/file/blob/...token.../
@@ -158,7 +158,7 @@ public class RegistrationBasicInfo {
          * @param registrationTime
          * @return
          */
-        public Builder registrationTime(Integer registrationTime) {
+        public Builder registrationTime(Long registrationTime) {
             this.registrationTime = registrationTime;
             return this;
         }
